Move TransactionDate class
