+++ vietnam_research/domain/valueobject/vietkabu.py
@@ -1,4 +1,8 @@
+import re
 from dataclasses import dataclass, field, InitVar
+from datetime import datetime
+
+from bs4 import Tag
 
 
 @dataclass
@@ -52,3 +56,34 @@ class Counting:
             return 0.0
 
         return float(s.replace(",", "").replace("%", ""))
+
+
+@dataclass
+class TransactionDate:
+    """
+    ウェブページから日付情報を抽出し、標準のdatetime形式に変換する。
+
+    Attributes:
+        th_tag: BeautifulSoupで抽出したHTMLのthタグ。
+    """
+
+    th_tag: Tag
+
+    def to_date(self) -> datetime:
+        """
+        BeautifulSoupで抽出したHTMLのthタグから文字列形式の日付('YYYY/MM/DD HH:MM')
+        を抽出し、それをdatetimeオブジェクトに変換する。
+
+        Returns:
+            対応するdatetimeオブジェクト。
+
+        Raises:
+            ValueError: thタグから抽出した日付文字列が 'YYYY/MM/DD HH:MM' 形式でない場合に発生。
+        """
+        raw_text = self.th_tag.find("b").text.strip()
+        extracted = re.search(r"\d{4}/\d{2}/\d{2} \d{2}:\d{2}", raw_text)
+        if not extracted:
+            raise ValueError("`YYYY/MM/DD HH:MM`形式が入力されていない")
+
+        date_str = f"{extracted.group()}:00"
+        return datetime.strptime(date_str, "%Y/%m/%d %H:%M:%S")
